requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
scipy==1.11.4

python-dotenv==1.0.1
redis==5.0.4
rq==1.16.2

boto3==1.34.162
botocore==1.34.162
